+++ include/driver/Logging.h
@@ -13,6 +13,9 @@ void CloseLogFile();
 tm TimeForLog();
 void LogFlush();
 
+// Returns true when enough time elapsed since the last log with this key.
+bool LogRateLimited(const char* key, unsigned intervalMs = 5000);
+
 inline FILE* LogOutput()
 {
 	return LogFile ? LogFile : stderr;